pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = "Learning management system"
authors = ["codewiztinsing <tinsingjobs2k@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.1"
djangorestframework = "3.13.1"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
django-split-settings = "^1.2.0"
pyyaml = "^6.0"
django-compressor = "^4.3.1"
django-crispy-forms = "^2.0"
crispy-tailwind = "^0.5.0"
django-braces = "1.15.0"
django-embed-video = "1.4.4"
django-ckeditor = "^6.5.1"
pillow = "^9.5.0"
redis = "4.3.4"
werkzeug = "2.2.2"
pyopenssl = "22.0.0"
cryptography = "38.0.4"
pymemcache = "^4.0.0"
channels = "3.0.5"
channels-redis = "3.4.1"
django-model-utils = ">=3.1.1"
django-allauth = "^0.54.0"
django-admin-interface = "^0.26.0"
django-widget-tweaks = "^1.5.0"
django-extensions = "^3.2.3"
django-bootstrap4 = "^23.4"
django-role-permissions = "^3.2.0"
django-roles-access = "^0.9.4"
django-crudbuilder = "^0.2.8"



[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "3.6.0"

[tool.flake8]
max-line-length = 79
exclude = '''
/(
    .eggs
    build
    dist
    venv
)/
'''

[tool.pre-commit]
repos = [
    "https://github.com/pre-commit/pre-commit-hooks",
    "https://github.com/psf/black",
    "https://github.com/timothycrosley/isort",
]

#<<<<<<< HEAD
#[tool.poetry.group.dev.dependencies]
#flake8 = "^6.0.0"
#pre-commit = "^3.3.1"

#=======
#>>>>>>> 5d54d47 (adding setupfiles)
#[build-system]
#requires = ["poetry-core"]
#build-backend = "poetry.core.masonry.api"
